requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-cloud-logging==3.5.0
cryptography==40.0.2
firebase-admin==6.1.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-secret-manager==2.16.1
pycryptodome==3.18.0
requests==2.31.0
python-dateutil~=2.8.2
Crypto~=1.4.1